core: support private data for mixer/enum/bytes control and widgets.
This patch copies private data to mixer/enum/bytes control and widget
elem which refers to the data. So the private data will output with its parent.

And three type-specific control lists are setup to replace the generic
control list for easy output:
- mixer_list		for mixer control list
- enum_list		for enum control list
- bytes_ext_list	for byte extended control list

The builder will write controls from the 3 control lists, with different
header types.

Signed-off-by: Mengdong Lin <mengdong.lin@intel.com>